Restructure benchmarks
