+++ crates/tako/benches/benchmarks/mod.rs
@@ -1,0 +1,2 @@
+pub mod core;
+pub mod scheduler;
